+++ fabric/src/main/java/com/lx862/mtrscripting/util/NetworkingUtil.java
@@ -1,5 +1,6 @@
 package com.lx862.mtrscripting.util;
 
+import com.lx862.jcm.mod.JCMClient;
 import org.apache.commons.io.IOUtils;
 import com.lx862.mtrscripting.lib.org.mozilla.javascript.NativeObject;
 
@@ -48,6 +49,9 @@ public class NetworkingUtil {
     private static void processRequestObject(NativeObject requestObject, HttpURLConnection connection) throws IOException {
         String body = null;
 
+        // Default UA
+        connection.setRequestProperty("User-Agent", USER_AGENT_STRING);
+
         if(requestObject != null) {
             if(requestObject.containsKey("method")) {
                 connection.setRequestMethod((String)requestObject.get("method"));
@@ -82,8 +86,10 @@ public class NetworkingUtil {
             }
         }
 
-        // Override User-Agent
-        connection.setRequestProperty("User-Agent", USER_AGENT_STRING);
+        // Override User-Agent if scripting restrictions not disabled
+        if(!JCMClient.getConfig().disableScriptingRestriction) {
+            connection.setRequestProperty("User-Agent", USER_AGENT_STRING);
+        }
 
         if(body != null) {
             byte[] bodyByte = body.getBytes();